Extract proof pack API module
