+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -111,3 +111,4 @@
 | LWB-R104 | Report batch operations action hook extraction | Hardened | `report-batch-operations-panel.tsx` owned report-batch materialization, status refresh, bounded run-once, archived-document lookup, pending state, errors, batch identifiers, and terminal-status gating inline beside table and archive rendering. That kept Gateway-only reporting/archive orchestration coupled to presentation and made fail-closed behavior harder to test without rendering the full panel. | Added `use-report-batch-operations-actions.ts` to own Gateway-only report batch creation, source-owned status refresh, bounded worker pass, archive metadata lookup, pending/error state, derived batch id, terminal run gating, and archive reference state. The panel now composes actions, status, item table, and archive summary while delegating API orchestration to the hook. Added hook coverage for call mapping, status refresh, run-once refresh, archive-reference trimming, terminal gating, and no-call behavior when required identifiers are absent. | Continue reducing reporting surfaces only where it separates Gateway orchestration from display composition or removes duplicated report/archive state. Keep batch membership, readiness, idempotency, report generation state, archive truth, retention posture, execution, delivery, and communication workflows source-owned and Gateway-backed. | `src/features/workbench/use-report-batch-operations-actions.ts`, `src/features/workbench/components/report-batch-operations-panel.tsx`, `tests/unit/use-report-batch-operations-actions.test.ts`; validation via focused report batch tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R105 | PM operating-quality review-action control extraction | Hardened | `pm-operating-quality-review-actions-card.tsx` mixed the bounded supervisory review-action form, preview/create readiness, create evidence rendering, ledger table, and detail evidence into one component after the review-action create slice. That made future control changes noisy and duplicated form-specific assertions inside the ledger/detail card test. | Added `pm-operating-quality-review-action-control.tsx` for the preview-before-create supervisory form, readiness strip, action buttons, and Manage create evidence. The review-actions card now delegates the command control and stays focused on ledger/detail rendering. Moved form behavior coverage into a focused control test while keeping a parent delegation test for composition. | Continue reducing PM operating-quality only where a slice separates command-control presentation from source-backed ledger/detail evidence or Gateway orchestration. Keep PM score calculation, fairness spread calculation, protected-class inference, PM ranking, HR/conduct decisions, client communication, approval, order, OMS, fills, settlement, and execution posture source-owned and Gateway-backed. | `src/features/workbench/components/pm-operating-quality-review-action-control.tsx`, `src/features/workbench/components/pm-operating-quality-review-actions-card.tsx`, `tests/unit/pm-operating-quality-review-action-control.test.tsx`, `tests/unit/pm-operating-quality-review-actions-card.test.tsx`; validation via focused PM-quality tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R106 | PM operating-quality API module extraction | Hardened | `src/features/workbench/api.ts` still owned the full PM operating-quality Gateway route family plus shared fetch, mutation, URL, and observability helpers. That kept a fast-growing BFF API barrel above 90KB and made PM-quality route changes harder to review without scanning unrelated portfolio, performance, risk, reporting, proof-pack, and construction helpers. | Added `api-client.ts` for shared Workbench Gateway fetch, mutation, URL, error, and bounded observability helpers. Added `pm-operating-quality-api.ts` for policy, score-run, support-summary, fairness-analysis, and review-action Gateway helpers, with `api.ts` re-exporting the same public API for compatibility. Kept all PM-quality request bodies Gateway-shaped and moved the duplicated fairness-analysis body construction into one helper without adding browser-side score, fairness, ranking, HR, client, order, OMS, or execution logic. | Continue splitting the Workbench API layer by coherent Gateway route family only when it reduces review blast radius. Preserve the public `api.ts` barrel until downstream imports are migrated deliberately. | `src/features/workbench/api-client.ts`, `src/features/workbench/pm-operating-quality-api.ts`, `src/features/workbench/api.ts`; validation via Workbench API tests, PM-quality hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
+| LWB-R107 | Proof-pack API module extraction | Hardened | After the shared API-client and PM-quality split, `src/features/workbench/api.ts` still owned the proof-pack generation, detail, Markdown, report-input, AI-evidence, and PM-memo Gateway helpers even though proof-pack UI orchestration already lives in `use-proof-pack-actions.ts`. That kept proof-pack mutation/request payloads coupled to unrelated Workbench API families. | Added `proof-pack-api.ts` for proof-pack Gateway helpers and caller headers, with `api.ts` re-exporting the same public functions for compatibility. Kept DPM wave proof-pack posture with the DPM wave surface, and preserved Gateway-shaped proof-pack payloads, idempotency, and bounded AI memo request outputs without adding browser-side proof-pack construction, hash generation, Markdown synthesis, prompt construction, approval, client communication, order, OMS, fills, settlement, or execution behavior. | Continue splitting `api.ts` by route family only where it removes real review coupling. Migrate direct feature imports away from the compatibility barrel in a later mechanical slice if it can be done without expanding behavior. | `src/features/workbench/proof-pack-api.ts`, `src/features/workbench/api.ts`; validation via Workbench API proof-pack tests, proof-pack hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
